# docs/openapi/swagger.yaml
openapi: 3.0.3
info:
  title: Azure DevOps to GitHub Migration Operator API
  description: |
    API for the Azure DevOps to GitHub Migration Operator, which facilitates the migration of resources from Azure DevOps to GitHub.
    
    This API provides endpoints for:
    - Discovering Azure DevOps resources (organizations, projects, repositories, work items, pipelines)
    - Creating and managing migrations
    - Monitoring migration progress
    - Converting Azure DevOps pipelines to GitHub Actions workflows
  version: 1.0.0
  contact:
    name: ADO to Git Migration Team
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /api/v1
    description: API base path
...
paths:
  /discovery/organizations:
    get:
      tags:
        - Discovery
      summary: Discover Azure DevOps organizations
      description: Retrieves a list of Azure DevOps organizations accessible to the authenticated user
      parameters:
        - name: client_id
          in: query
          description: Azure Service Principal Client ID
          required: true
          schema:
            type: string
            format: uuid
        - name: client_secret
          in: query
          description: Azure Service Principal Client Secret
          required: true
          schema:
            type: string
        - name: tenant_id
          in: query
          description: Azure Tenant ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Organization'
                  count:
                    type: integer
                    description: Total number of organizations
        '400':
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /discovery/projects:
    get:
      tags:
        - Discovery
      summary: Discover Azure DevOps projects
      description: Retrieves a list of projects in an Azure DevOps organization
      parameters:
        - name: organization
          in: query
          description: Azure DevOps organization name
          required: true
          schema:
            type: string
        - name: client_id
          in: query
          description: Azure Service Principal Client ID
          required: true
          schema:
            type: string
            format: uuid
        - name: client_secret
          in: query
          description: Azure Service Principal Client Secret
          required: true
          schema:
            type: string
        - name: tenant_id
          in: query
          description: Azure Tenant ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  count:
                    type: integer
                    description: Total number of projects
                  organization:
                    type: string
                    description: Organization name
        '400':
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /discovery/repositories:
    get:
      tags:
        - Discovery
      summary: Discover Azure DevOps repositories
      description: Retrieves a list of repositories in an Azure DevOps project
      parameters:
        - name: organization
          in: query
          description: Azure DevOps organization name
          required: true
          schema:
            type: string
        - name: project
          in: query
          description: Azure DevOps project name
          required: true
          schema:
            type: string
        - name: client_id
          in: query
          description: Azure Service Principal Client ID
          required: true
          schema:
            type: string
            format: uuid
        - name: client_secret
          in: query
          description: Azure Service Principal Client Secret
          required: true
          schema:
            type: string
        - name: tenant_id
          in: query
          description: Azure Tenant ID
          required: true
          schema:
            type: string
            format: uuid
        - name: include_empty
          in: query
          description: Include empty repositories
          required: false
          schema:
            type: boolean
            default: false
        - name: include_disabled
          in: query
          description: Include disabled repositories
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  repositories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Repository'
                  count:
                    type: integer
                    description: Total number of repositories
                  organization:
                    type: string
                    description: Organization name
                  project:
                    type: string
                    description: Project name
        '400':
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /discovery/workitems:
    get:
      tags:
        - Discovery
      summary: Discover Azure DevOps work items
      description: Retrieves a list of work items in an Azure DevOps project
      parameters:
        - name: organization
          in: query
          description: Azure DevOps organization name
          required: true
          schema:
            type: string
        - name: project
          in: query
          description: Azure DevOps project name
          required: true
          schema:
            type: string
        - name: client_id
          in: query
          description: Azure Service Principal Client ID
          required: true
          schema:
            type: string
            format: uuid
        - name: client_secret
          in: query
          description: Azure Service Principal Client Secret
          required: true
          schema:
            type: string
        - name: tenant_id
          in: query
          description: Azure Tenant ID
          required: true
          schema:
            type: string
            format: uuid
        - name: types
          in: query
          description: Comma-separated work item types
          required: false
          schema:
            type: string
        - name: states
          in: query
          description: Comma-separated work item states
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of items to return
          required: false
          schema:
            type: integer
            default: 1000
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  workItems:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkItem'
                  count:
                    type: integer
                    description: Total number of work items
                  organization:
                    type: string
                    description: Organization name
                  project:
                    type: string
                    description: Project name
        '400':
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /migrations:
    get:
      tags:
        - Migrations
      summary: List migrations
      description: Retrieves a list of all migration jobs
      parameters:
        - name: namespace
          in: query
          description: Kubernetes namespace
          required: false
          schema:
            type: string
            default: default
        - name: phase
          in: query
          description: Filter by migration phase
          required: false
          schema:
            type: string
            enum: [Pending, Validating, Running, Completed, Failed, Cancelled, Paused]
        - name: type
          in: query
          description: Filter by migration type
          required: false
          schema:
            type: string
            enum: [repository, workitems, pipelines, all]
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  migrations:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdoToGitMigration'
                  count:
                    type: integer
                    description: Total number of migrations
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Migrations
      summary: Create migration
      description: Creates a new migration job
      requestBody:
        description: Migration job configuration
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdoToGitMigration'
      responses:
        '201':
          description: Migration created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdoToGitMigration'
        '400':
          description: Bad request - invalid configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /migrations/{id}:
    get:
      tags:
        - Migrations
      summary: Get migration
      description: Retrieves details of a specific migration job
      parameters:
        - name: id
          in: path
          description: Migration job name
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          description: Kubernetes namespace
          required: false
          schema:
            type: string
            default: default
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdoToGitMigration'
        '404':
          description: Migration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Migrations
      summary: Delete migration
      description: Deletes a specific migration job
      parameters:
        - name: id
          in: path
          description: Migration job name
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          description: Kubernetes namespace
          required: false
          schema:
            type: string
            default: default
      responses:
        '204':
          description: Migration deleted successfully
        '404':
          description: Migration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
        status:
          type: integer
          description: HTTP status code
    Organization:
      type: object
      properties:
        id:
          type: string
          description: Organization ID
        name:
          type: string
          description: Organization name
        url:
          type: string
          format: uri
          description: Organization URL
        description:
          type: string
          description: Organization description
        region:
          type: string
          description: Azure region
        owner:
          type: string
          description: Organization owner
        createdDate:
          type: string
          format: date-time
          description: Creation date
        lastModifiedDate:
          type: string
          format: date-time
          description: Last modification date
      required:
        - id
        - name
        - url
...